Change latency distribution to a mapping of percentiles to latencies
Previously, an array of values was printed in the JSON output. But, the JSON
output is hard to interpret without the percentage values and depends heavily
on the order and exact values defined by the `olly gc-stats` tool. This commit
makes the JSON output more robust to changes in the percentile values used.